Handle NaN aggregation results when we have no results
When we have no results from our search and aggregate, we end
up with a NaN value. This breaks the query. This change checks to
see if we have NaN, and that also we have no hits, and zeroes this
result. Cases where we have NaN but have hits will still break the
query.